nanoops: align last 2 outliers to the bf16-load + dtype=fp32 sum pattern
Round-7 cleanup found two spots that the previous consistency pass
missed:

(a) `_fused_add_norm_fwd_kernel`: nw load still had a redundant
    `.to(tl.float32)` cast. y_f32 (fp32) * nw (bf16) auto-promotes nw
    to fp32 anyway; same fp32 product, one less cast.
(b) `_fused_add_norm_bwd_inline_kernel`: dnw_partial sum missed the
    explicit `dtype=tl.float32` kwarg. Input is fp32 here (HAS_NW=True
    branch with auto-promoted y_norm), so default-acc behavior was
    already fp32 — but the explicit kwarg matches the four other
    `tl.sum` calls in the file.

All four RMSNorm kernels (fwd, bwd_inline, inner, bwd) are now
identical in style: bf16 inputs loaded native, fp32 accumulators
declared via `dtype=`. 22/22 parity tests pass.

+++ nanoops/triton_fused_add_norm.py
@@ -211,7 +211,7 @@ if _HAS_TRITON:
 
         y_f32 = summed * rms_inv[:, None]
         if HAS_NW:
-            nw = tl.load(nw_ptr + cols, mask=col_mask, other=0.0).to(tl.float32)
+            nw = tl.load(nw_ptr + cols, mask=col_mask, other=0.0)
             y_f32 = y_f32 * nw[None, :]
         tl.store(y_ptr + offs, y_f32.to(y_ptr.dtype.element_ty), mask=mask_2d)
 
@@ -294,7 +294,7 @@ if _HAS_TRITON:
         )
 
         if HAS_NW:
-            dnw_partial = tl.sum(dy_t * y_norm, axis=0)
+            dnw_partial = tl.sum(dy_t * y_norm, axis=0, dtype=tl.float32)
             dnw_p_ptrs = dnw_partial_ptr + pid_m * D + cols
             tl.store(
                 dnw_p_ptrs,
